requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
openpyxl>=3.1.0
rapidfuzz>=3.5.0
pyyaml>=6.0
numbers-parser>=4.10.0
